backend/requirements.txt
Flask
pymongo[srv]
python-dotenv
openai
certifi
requests
Flask-CORS
dnspython